Extract the reviews and demos table.
